accept distance as float

+++ backend/src/main/java/com/Swoop/Swoop/controllers/RiderController.java
@@ -23,7 +23,7 @@ public class RiderController {
     @CrossOrigin(origins = "http://localhost:3000")
     @GetMapping("/addRide")
     public void addRide(@RequestParam String email, @RequestParam String start,
-							@RequestParam String end, @RequestParam int distance) {
+							@RequestParam String end, @RequestParam float distance) {
 		User u = userService.getUserByEmail(email).get();
 		userService.addRide(u, start, end, distance);
     }
